Adding thread debugging
This has somme implications:

a) endTrace is now a flag in Delegation

b) We introduce a lock to handle tracking of threads.

c) We also have now an id for each thread

d) We cannot have a unique Tokenizer for all segmentations.

When a thread is traced, it first enables a lock, which protects the access to the debugging instruction from interfereing with other threads.

Unlocking is the last action.

We also manage a master_lisp, since each thread has its LispE pointer.