input behaviour
input is now based on getchar. We have extracted the getchar from jag.cxx et built a jagget class to specialize it and avoid to large a code simply for get.

We have also implemented a much more powerful input, which can take an input that can be modifed.lispe.vcxproj.
We have also added a reloading i n the editor with ctrl-l.