splitline preparation for windows
First, we now have a special_characters UTF8 manager in main.cxx. We do not rely on lispe anymore.

Second, we now implement a faster version of size_upto that do not require a removal of all emojis.

We will update the Windows version whenever we have access to one.