Massive modification
Now, we implement a new kind of structure to handle lists: LIST and ITEM. The main improvement here, is that a cdr instruction shares the structures with the element it takes its cdr from. The new list from cdr is built around the same ITEM object, with a 'home' value to indicate where it starts. The element in the list are cleaned only when no other sublists is alive. Everytime an ITEM is shared, its status is increased by 1, when a LIST pointing to this ITEM is freed, then this status is decreased. Only when this status is set to 0 is the list cleaned.