New features
a) We add regular expressions to the editor.

b) We add color highlighting for []

c) We add [] as a way to implement LispE code to have some visual cues on specific structures such as pattern functions

d) We introduce checkInteger that returns either a value if the object is a number or throw an exception otherwise.

e) We redesign operation functions to handle list.

f) we add sum and product

g) flip can flip a simple list now

h) We create a copy of 'e' in thread_store out of lock

i) darkmode can be called from the command line.