Improvements
a) Instruction methods are now accessed through an array and no longer through a map. We initialize this array woth eval_error for values for which there are no evaluation method.

We have also got rid of some if through booleans_ in comparison methods.

b) shuffle is now implemented with a simple loop and a swap... We use a random generator to find the slot to swap with.

c) We have implemented a introsort with a greedy version for small arrays.

This version sorts values 20% faster.