Minor corrections
a) The throw creates its own Error object in zip and zipwith.

b) Correction on the editor for Windows. In particular, there was a mouse string that would be displayed in the console.

c) We count the number of clicks on Windows in a different way.

d) minizork_fr and minizork_en