Corrections
In case of error in an if, ife, check or ncheck, an error might have posed some problems on releasing the test in firt_element. We use _releasing to clean first_element and set it to nil.

We also set the terminal flag to the last element of a block.