Modifications on the reader
First, we introduce Dicationary_as_buffer, which is used to parse data dictionaries.
A data dictionary is introduced with @{..}. While a {..} creates the instruction to creaet a dictionary, @{...} creates the dictionary directly.
We also modify the JSON reader to use this new dictionary object to create dictionaries out of JSON files.

Second, we introduce the "multiloop", which allows for parallel looping on different lists.