Two modifications
a) We introduce newInstance to handle operations between a scalar and a list. In that case, we transform the scalar into a copy of that list with this scalar as being the only value for this list.

b) Our shape is now tested for lists, in order to check if all elements match the first element shapes in depth.