Bug corrections
lst was not instantiated when computing its length in product and sum.
We also add some new characters such as pi etc in conjunction with _pi etc.

We also correct a bug that prevented a correct parenthesis with a large UTF8 character in the input. We now convert to wstring before computing the parenthesis position.