New improved version
This new version is much faster than the previous one as it replaces unordered_map in stacks with a home-made versions (binHash in mapbin.h).
We have also replaced wstring with u_ustring, which is actually based on std::u32string.
This allows for a much simpler handling of strings across different platforms.
The problem is that on windows wstring is encoded in UTF16, while it is encoded in UTF32 on a ll other platforms. This new solution provides a consistent implementation of strings across all these platforms.
We have also got rid of provideFloat and provideInteger, which were a bit too slow.