Bug and improvements
First, we need to record functions at compile time, otherwise they are not available for defpat argument evaluations.
Second, we now have a better way to handle different objects, such as numbers, integers etc..., which are stored in a pool.
We have some specific derivations (Numberpool, Integerpool), which when their referenc counter (their status) is decremented, are now stored into a pool, insted of being deleted.
Finally, on Mac OS, we now compile with the flag: -Ofast for better perfomances.