Massive Update
First, we have improved the speed of the language by a factor of 2, thanks to many cache mechanism.
For instance, values such as strings, integers or numbers are now stored in pools instead of being deleted when released.
The same applies to lists, sets  and dictionaries.
We have simplified the reference counting mechanism. there is no top flag anymore.
The top flag was used mainly in dictionaries to handle the storage of an element. We only increment it once (or decrement it twice)
We have also integrated in debug mode a mechanism to track the lifecycle of all objects in the language to detect potential memory leaks.
We have corrected some major bugs in the eval function, when many objects were not correctly released after the evaluation. In particular, there was an issueif a function was created through eval.
The stack has been simplified. We do not store constant object in it, this is the only case when an object is not stored in the stack.
The status is now an uint16_t. s_constant and s_protect are now stored as bits in this status.
is_portected is now used everywhere in order to detect if a value is a constant.
We have added a type: float and floats for float values on 32 bits.
We have also added many new functions:
stringf to display a number with a specific format (as sprintf)
takelist, droplist, which are now also available in conjunction with l_filterlist and l_maplist.
infix can now take an argument that is evaluated during the execution of a program.
addr_ returns the actual pointer address of an object
mloop to loop in multiple containers at once
lloop to loop in multiple lists at once
setrange, which is used to replace a range of values in a list with a new value
sign modifies the sign of a number
signp tests the sign of a number

We have also replaced many vector declarations in various methods with a call to provideStrings or provideIntegers, to reduce the number of vector creations on the stack.

We have fix many problems in the ways elements were released when an error would occur.

We also generalized the use of u_ustring everywhere and in all libraries.
We also added a binSet, which is only store the uint16_t ids..
In vecte, we now use malloc instead of new to be able to use realloc, but also to compatible with C if necesasry.
We did the same in ITEM (listes.h)