Fixing some bugs and memory leaks
First, the debug elements to handle memory leaks are now called __idx , __lc and __indexes, to avoid some potential confusion with actual variables.
Second, we have replaced the specific List_plus, List_minus etc... with. a generic call to List_execute and List_basic_execute. These objects are created with their specific method call embedded.
Third, there was a bug in extract and replace_in, when the first element was an integer and the following one a string. In this case firstisString should be 0. furthermore in the same function, the indexes were not released.
Fourth, abstractSyntaxTree takes a quoting parameter to take into account a quote applying to a recursive description.
Fifth, in Atom::unification, record is now the prevalent test that subdue any other considerations. The previous case would prevent for instance an atom to be recorded with its own value.
Sixth, check_quote takes as argument the position of the element for which we want the transformation. We can now scan a structure and replace each quoted element with a Listargumentquote element.