Bug corrections and improvement
First, we have corrected some more memory leaks, when pattern rules are applied against sets in particular.
Second, we have simplified and improved the pattern rule compiler, which was a little bit too bloated.
It is much more streamlined and consistent now. In particular we now accept that variables are not the last element of the call.