Bug correction on Xcode gui
First, we add a "running" flag to check if a LispE program is running.
Second, in syntax highlighting, the "limit" is now computed as the number of bytes in a UTF8 string.
In some cases, the end of the text was not properly isolated, since the size was based on the number of characters not the number of bytes.