Different improvements
First, we update the debugger, which in some cases would skip lines.
Second, we now have a specific template: vecte_a (and vecte_n for strings), which provides the same functionality as LIST. This makes possible the implementation of a cdr for Strings, Numbers, Floats and Integers that shares the same array, without any need for duplication.
Third, Pattern Functions can now accept description with the Kleene operators.

It is now possible to declare:

(defpat ( ["<" a+ ">]"] ) etc...

where a will be a list of all values between "<..>".