Bugs corrections with new features
First, we provide now: llist, which creates a linked list (class LList)
The current list implementation is based on an array.
Second: ;; can be used to comment a few lines at a time.
The first line to be commented must contain: ;;
and the last line: ;; as well

We also provide now: popfirst, poplast and pushfirst and pushlast, which are complement (and sometimes redundant) to push and pop.

We add minmax, which returnsa list containing  the minimum and the maximum elements from a list.

There was a few bugs:

replicate has been fixed.

The resize function in some templates would not use the new size as a starting point to compute a new array size.

if (t >= sz) --> we need sz = t << 1; and not sz = sz << 1;

Second, the list is implemented in a way that only the last list to be cleared will cleared all elements. It reduces the number of list traversal. When a cdr is applied the list is no longer traversed to increment each element. Elements status is only incremented once when the element is inserted in the list.

We also corrected a bug in push. If second_element is the result of a cdr, we cannot released after adding its copy to the main list, as it risks decrementing this element.
In maths.cxx, some elements were not properly released after an operation.