Bug correction and new feature
First, the way cycling lists were "cdr" was wrong. We need to mark the first element of that new cdr to start looping, not the current head.

Second, cdr can now indefinitely cycle. A succession of cdr on a linked list that contains cycle will loop for ever. However, the method "loop" on a linked list does not loop for ever and stop after one traversal.