A bug and improvements
First, there was a bug when accessing a linked list through index. We did not update the code for it=it->next(), it was still it++.
Second, we have improved the way lists are computed when an operator is applied to them. We have in particular integrated the linked list as special case.
Third, we got rid of liste.size() comparison in value_on_index and protected_index for LList. liste.size() is very costly as your need to traverse the whole list to know its size.