Bug correction and improvement
First, the way the last function was implemented had an issue. We need to reset the marks again at the end of the process or the next _previous()_ will fail since it will share the same mark with the previous element.

Second, we got rid of all calls to liste.size() as much as possible for LList elements. In a linked list the size is recomputed each time by traversing the list.