Different improvements
First, we correct a bug in lispe_gui.cxx, we did not test if function was NULL, which crashed some of our examples.

Second, we generalize reserve for most lists.

Third, there were some issues with some methods for LList.
For instance, when using at or b_at etc., we need to reinitialize the inner mark to be able to loop again.
Finally, in a loop, we need to use the push_front version at LList level to be sure that element reference counter is taken into account.

The rho method can now detect the size of the element that is taken into account to create the different sublists.
Finally, irange has now the same argument order as range when used with 3 arguments.