Redesigning rho
Now, rho takes into account the type of list that is given as input. In this way it is now possible to creates embedded linked list objects:
(rho 2 3 4 (llist 10 20 30))
Note that llist is built backwards.
We also add: emptyp, which checks if a list is empty. It is much faster especially for linked lists than checking the size.