New features and bug corrections
New type: seti to handle sets of integers...

Instructions:
replaceall, which replaces in situ a value in a container. Returns the number of modifications

rank is now extended to regular lists as long as they contain regular size elements.
We also have irank to loop on columns in a matrix.

We also add the function: count, that counts the presence of a value in a container.

Bug corrections:

bit_not, the value was modified in situ.
reduce: the use of a list to handle the elements in another list were wrongly handled.