New features and bug correction
There was some cases of memory leaks in nconc and zip due to the use of value_on_index in an append. If the value is pushed into a value vector, then the value can be duplicated but not released.
There was also a case of memory leaks in loops. We need to release e before setting it up to null.
In evall_zipwith, there was an issue to release the container in the catch.

We have add three major operators: &&&, ||| and ^^^ that apply to lists and sets.
&&& plies an intersection between two lists
||| applies a union
^^^ applies an exclusive union.

The inner product can now apply to regular lists. There is now version of combine for regular lists.