Correction and new features
First, we now offer a set type, which can be used to store objects. There is also a new sets type to store strings in sets.

The sort algorithm could not handle lists of lists. This is now fixed. We push in our comparator a quoted version of the object to make the comparison possible between lists.

There was an issue when an operator was applied to an empty value list.
Furthermore, we now return an error when an operator (except for +) is apply to a list of strings.

maplist, droplist, filterlist and takelist can now apply to any container. We have extended the notion of begin_iter to all containers.

The mark to handle cycles in linked list is now an uint32_t. This simplifies the ways this mark was used, as now the value will loop among 4 billion values.