Missing elements in Heap class
We now provide some of the overloaded methods that were necessary to use the heap as a proper container.

Missing:
 Boolean(), which returns false when the heap is empty.
 isEmpty()
unify, which is used in pattern matching functions.
heap_ is now also a primitive in data structures for matching.
The replace methods for lists can now takes a negative index.
We corrected some issues in convert_in_base. For instance when base==1, we created an infinite loop.