Two corrections
First, we did not "quote" the first element in a call to scan or backscan with a lambda expression.

Second, we fix the debugger, which did not take into account the fact that specific calls to functions could be replaced with a specific object.
We introduce a trace_on variable, which works in tandem with trace. When trace is debug_none, trace_on can still be on.