Minor and major modifications
First, a minor modification is in the way types are now normalized across all libraries. A type is now: "name_".
The final "_" is now normalized to define a type atom.

Second, a major improvement is the fact that copyatom takes a lisp object as argument. Thanks to this, many objects can now use object pools. This result in some interesting speed and memory improvements.

We now provides dictionary, dictionary_n and dictionary_i to create dictionaries. This is more consistent with way all other objects are created and push "key, keyi and keyn" in a different utilization perspective. These new functions create dictionaries in same way as other types such set, integers or list.

l_emptylist is now available, which takes into account the evaluation of the empty list. This was a bit of a weird mistake in the interpreter as empty list would trigger an error that would be dismissed in eval_error. This resulted into a massive speed improvement.

We have also improvement the way elements are checked in a vecte object. The same improvement went into vecte_ and vecte_n, List).
Basically, we removed the test from the loop.

We now have specific methods to push elements in different containers:
push_element_first, push_element_last and push_element.
The eval_push, eval_pushfirst, eval_pushlast are now very simple.

We have removed as much as we could try/catch from evals methods.
Many of them did not make any sense, since no elements needed to be released in case of failure.

We also corrected a bug in argumentsize. The list size was not taken into account to explore parameters. When zero it would crash the interpreter.

lispe_curl also now has overloaded methods for protected_index and extraction, in order to be used as strings as output by the curl_url method.