Adding file_getstruct to the instruction set
file_getstruct is used to read successive structures from a file, which starts with a specific character and ends with another.

(file_struct stream "{" "}")

The file can contain sub-structures of same sort, which means that that the method will only stop when all sub-structures have been consumed.
Hence, the method can return {{a:b} {c:d}}.
The method can be called as many times as necessary to read all balanced structures. When the last structure has been read, the system returns nil.