Speed improvements
First, we now handle sums and products of number vectors in vecte, which is closer to the metal than the previous methods.
Second, we now have direct access to plus, divide, multiply and minus through plus2, plus3 etc..., which are compiled on the fly according to their number of parameters. We have now direct classes for these operations (as it was the case for power2):
List_plus2, List_plus3, List_plusn etc.

In dictionaries, we get rid of the try/catch and use an iterator based on find. In this way, we can both test and return or modify a specific value at a specific key. The speed improvement is really interesting about 2 to 5 times faster.

duplicate_constant now takes a lisp as argument, which makes possible the use of object pools. This method is called when pushing an element into the stack.

Some refinements in isequal for value vectors.

decrement, increment do not use a if anymore but use not_protected() return value to increment or decrement.