Editor improvements and rank
First, we finally come up with the final version of rank. This version corrects the problems we had when the size of the list was 2.
We now have the same buildList method, which called for Tensor and List, which build a list of indexes backward to access each element.

Second, we corrected much of the problems that plagued the LispE editor with long strings or long comments.
We now have a parallel list in which long lines are marked with a specific character.
In coloringline, we now check this specific character before hand to assess if the line has been marked as a long string or a long comment.