Bug correction
First, we correct a bug that prevented expressions such as: (+= (@ x 1)...) to be evaluated...

Furthermore, in these expressions such as (+= x 10) or (+= (@ v 1) 10), we check them at compile time to select the right method (only for divide, plus, multiply and minus)

We also get rid of the List_basic_execute class that has become redundant.

Finally, we correct a bug in configure.py. The VTERM_MOUSE option was not correctly written into Makefile.in (missing CR)

We also add an APPLE flag to Makefile.in for mac os compiling.