Bug in bit_not
There was a bug with bitnot, first_element should be tested against e before being released.
The result was a mess where a value would put back into the pool and stored in the stack.

We finally get rid of the if_protected test with a single minus in remaining decrement methods.