Jag Editor bug corrections
There was a problem when deleting characters in a long line. The lines were not correctly re-displayed on screen.
In particular, prinline now takes into account the fact that the line number should not be displayed for next segments of a long line.

Finally, there was an issue when launching the terminal with the no-prefix option (line numbers are not displayed).

We provide a new command on the command line of lispe, which is: -mg.
This command makes it possible to modify the right margin on screen.