Massive Normalization in code
First, all atomic values (numerical values) have the same inner field name: content.

Second, Chaine_UTF8 is now UT8_Handler.

Third, we now have a new class: Quoted, which is used to put in "quotes" arguments to functions that were passed as arguments.
We get rid of the previous mechanism, which was to have a specific variable in List that would indicate that the list should not be evaluated.
This new class makes it possible to use any arguments that will not be evaluated such as lists and atoms. Also, it protects elements from being deleted  in a call as they are incremented when pushed into a Quoted object. Finally, the simplify the release of a "call" variable, which will simply release a Quoted element. Finally, we also introduce a in_quote method that stores an argument in such a Quoted object.

Fourth, the "filp" method is now much simpler. We introduce a generic swap method, which helps merging all flips on a lists to one single piece of code.

Fifth, the method "insert" can take a function or an operator as last argument, so that elements that are inserted can automatically be ordered.

Sixth, there was normalization on the use of protected_index and value_on_index.  Now every protected_index returns an element from a dictionary or a list without duplicating it. The only exception are value containers, in which case we still need to return a new object.

Seventh, we have generalized the use of not_protected to decrement the status of an object, removing a "if" from the different decrement methods.

Eighth, we have normalized calls everywhere in eval.cxx with the use of provideCall, which creates a list in which arguments are Quoted objects.
In particular, we have now an eval_body_as_argument_min, which provides a direct call to functions and operators.

Tenth, we have redesign "apply", which is now much more efficient, as we now detect if it is a List object, and only insert the function call if necessary. For this purpose, we have also redesign asList, that now takes a list as argument to push the element of the current value container.

Eleventh, the call to comparison in sort takes now advantage of the new Quoted objects. We can now easily compare lists together.

Finally, UTF8_Handler can be shared with a LispE object, as the constructor can take a UTF8_Handler pointer as argument. This reduces the number of these objects in memory to one in most cases.