Emojis and const
a) First , we put whenever we can a const in loops such as:
 for (auto&...)

which are now:

for (const auto&...)

This has some benefits for the compiler to generate a more efficient code.

b) We introduce a new way to handle emojis through an automaton. We have downloaded from https://unicode.org/Public/emoji/14.0/

This class handles the detection of full sequences of emojis characters for UTF8, UTF16 and UTF32.

It replaces in the code the previous version, which was based on emoji 12.

c) We have modified the way the color highlighting was implemented in the Mac OS gui.
We now extract the UTF16 internal string, which is used for parsing. Since, the string has the same format as the internal mac os one, the indexes are perfectly aligned.