Improvement on jag
It is now possible to modify the syntactic colours from the command line.
We add different commands within LispE that can be used to modify the syntactic colours: colors and syntax.
syntax syncolor can generate a command line option, which can be used in conjunction with LispE and jag: -syncolor ....

We also correct a bug in the debugger. In some cases, lisp->getvalue can return NULL, which needed to be tested.