Modifications on getMethod
We get rid of the different variables, which were used to check the states before executing an instruction. We replace these variables with an actual test.

We get rid of getMethod, which is now replaced by some more specific tests in eval_pattern.

We also introduce an Argumentdictionary type to handle dictionary unification, which now has a specific case for Dictionaryi type.