Speed improvement
We have simplified the way lambda are executed in maplist, zipwith and heaps. We now use a specific sets of instructions: record_or_replace and reset_in_stack to handle the management of local variables.
The main idea here is that we do not want to create a specific stack element for lambdas as they can use external variables for their purpose. However, some lambda variables can have the same name as external variables and in that case we need to keep track of these variables to reinstate them once the lambda execution is finished.
We also need to keep their status in order to contravene any modification of these external values during the execution of a loop in map/drop/take/filter list or zipwith.