Some modifications
a) We add a parameter to maplist and zipwith. By default these two methods return a list of values based on the first element that is computed. We add a new parameter that forces this output to be a regular list.

b) We also correct a bug in the way pattern functions are indexed. We no longer index these functions on their first element if this first element is a list of lists.

c) We add the _digitp_ method, which was missing from strings.

d) prettify has a new parameter to define the maximum size of a line on screen.

e) We also correct a bug in the tokenize rule section, when a rule starts with '%'

f) We now handle _return_ with two specialized classes.