Two bugs
First, when we use to_list on a list, we return the list itself. However, we need first to get rid of the result list, and second we need to check if values and results are different before releasing values.

Second, when abstractSyntaxTree returns an empty list, we replace it with the EMPTYLIST object.