Bug in Internal Debugger
Breakpoints in files that were loaded from within the main file, were not taken into account.
id_file sets now a temporary id for a file that will be loaded later. In this way, when the secondary file is loaded it already has an id (within allfiles) but still no entry point, which is our clue to actually load it. Otherwise, we simply return its entry point.