Different improvements
First, we update the error messages for car and cdr. We display more information on the rationale behind the error.

Second, we correct the way "atom_" definition is handled in pattern matching rules.
Actually, the issue stems from the fact that label() returns the atom value for an atom while most other elements returns their type by default. We introduce a new checkDataStructure, which takes an Element as input. If the atom was not used as an index, we then return its type. For all other elements, type_atom returns v_null except for atoms.

Finally, we introduce a notion of depth, which is computed while compiling the patterns in pattern functions (defpat). We use this depth to explore sublists up to a certain point.