Some modifications and improvements
First, nconc and nconcn compute now elements in their order of appearance.
Second, Pair is now a type of elements not a list. This simplifies the way (cons 'a 'b) is built.  Furthermore, it solves the issue of Linked Lists for these elements.
Third, we also introduce a new notation: '(a . b).
Fourth, we get rid of the pair parameter in duplicate_constant.
Fifth, cons and consb do not create any new list, we insert the elements at the first position in the list.
Finally, we modify the code for extraction and cdr to take this new Pair element in account. Note, that Pair also provides its own arithmetic instructions now.