Better quote management
We have now a better implementation of how quoted expressions are compiled. Instead of a complicated (and bogus) implementation that tried to keep the parsing iterative, we have now a recursive version, which is more reliable.
Also we now handle final element in lists such as '(a . b) in a better way with the use of conspoint, which creates a linked list and adds the final element in a specific way.
c) basically, we get rid of the Pair class, which proved to be an impasse.
d) The conspoint impacts the LList class, we have now a u_link_last class to handle final elements in lists.

e) We also push a new example: cfg.lisp, which is worth its salt.

Look at it in awe...