Adding space name for functions
We provide now three new ways to handle space names in LispE:

(defspace name ...), which defines a space name, in which functions declared in the body belong to.

(space name ...), functions that are defined in a space, can be called through space followed with name. The body is handled as a block.
; We declare a space name: truc
(defspace truc
   (defun calculus (x y)
      (+ x y)
   )
)
; We execute a block of instructions within this space
(space truc
    (setq r (calculus 10 20))
    (setq v (calculus 20 40))
)

The other modification is with load, which can take a second argument that defines a name space where the functions from the file will be stowed.

(load "test.lisp" truc)

Again truc is our space.

This space avoid the problem of having different functions with the same names.