Potential bug correction
We introduce binhash, which is has table based on int16_t keys. The other versions use a _base_ value to avoid creating too many intermediate tables in the map.

binhash does not use "base", which enables a slightly faster implementation, but also makes it more resilient in threads, since "base" might be modified in a thread...