Redesigning high level function composition
We have now a new way to compose high level functions (map, filter, takewhile etc.), which is based on an inner list: composition_stack.

Whenever we detect one of these functions, we push it into composition_stack. When we reach the end of the program or detect a different kind of functions, we compile it into "loop" structure. These functions are in fact macros.
The difference with the previous implementation is that we have the full stack now to compose, while we had to store each element individually and compile them at each stage.
The result is a much better and simpler way of composition, since the full stack gives access to the whole structure at once.

Second, we have also added a new "system atom": _root, which returns the list of instructions that corresponds to the program in memory.