Bigs and improvemetns
a) We could not compute the stack size in the case of lambdas, since lambdas use the local stack. We now have a depth_stack variable, which is incremented whenever a checkStack is performed. We use resetStack to decrement it.

b) We have added some direct call lists. We now have a direct call for the empty list.

c) There was a bug in the way List_if_eval was built. The terminal flag was not correctly set, which prevented tail recursion to work properly.

d) We have a better implementation of composing, with a lot of functions that are directly computed as direct calls.