Some important modifications
The most important modification is related to the new improve tokenizer, which is now compiled into an automaton.
The rules remain the same, however the way they are compiled and executed is completely new. For instance, we have now a vectorization, which is based on 1611 element table, to accommodate most of the European alphabets.
The execution has now been simplified and can accommodate most C++ types of strings, such as UTF8 strings, UTF16 strings and UTF32 strings.
The process has been divided into two different parts, on the one hand the rules themselves and the other hands the type of objects into which the rules output is stored.
The encoding is taken into account through different template functions.

Second, since LispE is now available as a WASM library, we had to compress the size of the stack.
We introduce a generalization of the mechanism that transforms a call to a function into a class .
Basically, we now create an instance of a class corresponding to an instruction (see straight_eval), whose eval is direct. We avoid the indirect calls to instruction through Delegation::evals and Listincode.
However, we still keep this implementation to be able to dynamically evaluate a list.

Third, we introduce complex numbers, which can be freely integrated into numerical expressions. See mandelbrot implementation in gui/examples for an example.

Fourth, we provide an example of how to use LispE as a WASM library within a browser.