Makefile update
a) Now all Makefiles share the same basic compiler.

b) On Mac OS, when linking the anaconda python library, we need to use the following flag: -undefined dynamic_lookup.
The reason is that the python dynamic library is actually static and already provide an initialiasation that conflicts with the pytamgu, creating a crash.