Matrix multiplication of vectors of numerical values
Matrix product is the implementation of the matrix multiplication with the operator: +*

This matrix multiplication takes as input vectors of numerical values with their shape.

(+* m1 '(shape11 shape12) m2 '(shape21 shape22) )

It returns a list that contains first the new shape and then a vector of values of the type of m1 and m2.