Minor modifications and a bug correction
Minor modifications
cab147c7
a) We redesign outer and inner to have their operators next to the function. The operators are now placed as for apply.

b) We modify rho to make the use of shape as a list possible

c) We correct _inner_. The return value of isPureList are now based on a enum, which was not taken into account.