Some massive updates
a) we introduce stringbyte and stringbytes, which record strings as UTF8 strings in memory.
b) we introduce "bytes", which returns the internal encoding of a string.
c) we also extend the notion of matrices and tensors to these new types.
d) reduce, scan have a fourth argument, to apply the operators to a slice of the current lists.
e) fread now returns a stringbyte.
f) We also introduce bread, bappend and bwrite, which handle a file as a list of bytes
g) cutlist is now slice (we still keep cutlist though)
h) we introduce a new notation for byte strings: b"my string"