Bug corrections and new instruction
a) We corrected a bug for the split instruction, which would only return a stringbyte.
b) We also corrected a bug where in case of error reduce and scan would release an element twice.
c) We add the mark instruction, which takes a mask of Boolean values to build a new list out of two other lists.