Bugs correction
a) There was some elements when creating threads that were not properly cleaned. We introduce ThreadElement to take care of these elements: the thid in particular.

b) When dealing with threadspace variables, we need to set the preparingthread Boolean, so that non value can be created within one of the pools.