Adding binary definition
We also modify the signature of execute_file. The variable path is now a copy.
Renaming of execute_some_code into execute_code.
l_lambda in eval.cxx returns itself. It is not necessary to quote them anymore.
We get rid of numbers in Segmentingtype. The conversion is done later at compile time.