Adding tail recursion
We also add label, that can add a global value to the environment.
We also correct a bug that prevented recursive calls.
Finally, we introduce a very simple but efficient tail recursion mechanism.
To sum up. We have now a stack on which we push our instructions, but also function calls. When a call is done to a function, we check if the top of the stack is the same as this function. In order to make the mechanism work, we have a few tricks. First, we need this call to be the last of an instruction set, second, if this call is in a cond or an if, then it is still considered as a potential tail call. Hence, when a call is done from within a cond, an if or an else, we clean the top of the stack in order to keep on the top the last function call.
Then, we have two more modifications. In a block, the last instruction is also called with the stack being cleaned from the l_block. The same when executing the instructions of a function.