Adding a new function and correcting a bug in defpred
defpred wasn't cleaning the stack properly when switching to a different function body.
Furthermore, in case of total failure, arguments wasn't released.

We also add: setqv, which works as setq but returns the value that was stored instead of true. This is very useful with defpred function to check if the value that was stored wasn't a fail.